# PKMVP-BE/Pkmvp.Api/Models/TokenResponse.cs
using System;

namespace Pkmvp.Api.Models
{
    public class TokenResponse
    {
        public string AccessToken { get; set; }
        public DateTime AccessTokenExpiresAt { get; set; }
        public string RefreshToken { get; set; }
        public DateTime RefreshTokenExpiresAt { get; set; }
    }
}